Updates to Class Concept Binding to support enums as well as ordinary classes.
