+++ common/src/main/java/dev/ikm/tinkar/common/bind/annotations/ConceptAnnotation.java
@@ -4,6 +4,11 @@ import dev.ikm.tinkar.common.bind.annotations.names.FullyQualifiedNames;
 import dev.ikm.tinkar.common.bind.annotations.names.RegularNames;
 import dev.ikm.tinkar.common.bind.annotations.publicid.PublicIdAnnotation;
 
+import java.lang.annotation.*;
+
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.FIELD})
 public @interface ConceptAnnotation {
     String name();
     PublicIdAnnotation publicIdAnnotation();
