+++ common/src/main/java/dev/ikm/tinkar/common/bind/annotations/names/FullyQualifiedName.java
@@ -23,7 +23,7 @@ import java.lang.annotation.*;
  */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE})
+@Target({ElementType.TYPE, ElementType.FIELD})
 @Repeatable(FullyQualifiedNames.class)
 public @interface FullyQualifiedName {
