+++ common/src/main/java/dev/ikm/tinkar/common/bind/annotations/names/FullyQualifiedNames.java
@@ -25,7 +25,7 @@ import java.lang.annotation.*;
  */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE})
+@Target({ElementType.TYPE, ElementType.FIELD})
 public @interface FullyQualifiedNames {
     FullyQualifiedName[] value();
 }
