+++ common/src/main/java/dev/ikm/tinkar/common/bind/annotations/names/RegularName.java
@@ -28,7 +28,7 @@ import java.lang.annotation.*;
  */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE})
+@Target({ElementType.TYPE, ElementType.FIELD})
 @Repeatable(RegularNames.class)
 public @interface RegularName {
