+++ common/src/main/java/dev/ikm/tinkar/common/bind/annotations/names/RegularNames.java
@@ -24,7 +24,7 @@ import java.lang.annotation.*;
  */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE})
+@Target({ElementType.TYPE, ElementType.FIELD})
 public @interface RegularNames {
     RegularName[] value();
 }
