+++ common/src/main/java/dev/ikm/tinkar/common/bind/annotations/publicid/PublicIdAnnotation.java
@@ -19,7 +19,7 @@ import java.lang.annotation.*;
  */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE})
+@Target({ElementType.TYPE, ElementType.FIELD})
 public @interface PublicIdAnnotation {
     UuidAnnotation[] value();
 }
